Waves 1 and 2 minor tweaks.
Bot now switches to javs after 4 hits on the mager with ven bow, if mager still alive.
Bot now starts on the correct tick wave 1 100% of the time.
Changed how movement works while bot is killing rangers on wave 2.
Bot now switches to ven bow if killing the mager, and the meleer is next to it.
Changed method of tick manipulation on core processor.